requirements.txt
# Core Libraries
numpy==1.21.6
pandas==1.3.5
scikit-learn==0.24.2
joblib==1.0.1

# Data Preprocessing for Imbalanced Datasets
imbalanced-learn==0.8.1

# Hyperparameter Optimization Libraries
scikit-optimize==0.9.0     # For Bayesian Optimization (BayesSearchCV)
tpot==0.11.7               # For Genetic Algorithm-based Optimization
tune-sklearn==0.4.4        # For Hyperband and advanced search with Ray
optuna==2.10.0             # For Optuna Optimization

# Metrics and Evaluation
scipy==1.7.3
